feat: 바꾼거 겁나 많음 나도 모르겠다 커밋 슛~~~

+++ backend/src/main/java/com/zzf/backend/domain/bank/repository/DepositRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepositRepository extends JpaRepository<Deposit, Integer> {
+public interface DepositRepository extends JpaRepository<Deposit, Long> {
 }
